# src/GestorIA.Api/openapi/v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "GestorIA API",
    "description": "SPEC-009. Money is a string with two decimals, dates are ISO-8601, errors are RFC 9457 problem+json.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/profiles": {
      "get": {
        "tags": [
          "profiles"
        ],
        "summary": "The stored profiles: none before the first is created, one after.",
        "operationId": "listProfiles",
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProfileView"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKey": [ ]
          }
        ]
      },
      "post": {
        "tags": [
          "profiles"
        ],
        "summary": "Stores the profile. Local mode keeps one: a second is refused with 409.",
        "operationId": "createProfile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileInputDocument"
              }
            }
          },
          "required": true
        },
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileView"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKey": [ ]
          }
        ]
      }
    },
    "/api/v1/profiles/{id}": {
      "get": {
        "tags": [
          "profiles"
        ],
        "summary": "One stored profile.",
        "operationId": "getProfile",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileView"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKey": [ ]
          }
        ]
      },
      "put": {
        "tags": [
          "profiles"
        ],
        "summary": "Replaces a stored profile with the one in the body.",
        "operationId": "replaceProfile",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileInputDocument"
              }
            }
          },
          "required": true
        },
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileView"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKey": [ ]
          }
        ]
      },
      "delete": {
        "tags": [
          "profiles"
        ],
        "summary": "Deletes the profile and everything stored for it. Nothing is kept: export it first to keep a copy.",
        "operationId": "deleteProfile",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKey": [ ]
          }
        ]
      }
    },
    "/api/v1/profiles/{id}/export": {
      "get": {
        "tags": [
          "profiles"
        ],
        "summary": "Everything stored for the profile, in one versioned document (SPEC-009 §2.1). Personal financial data: sent with Cache-Control: no-store.",
        "operationId": "exportProfile",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileExport"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKey": [ ]
          }
        ]
      }
    },
    "/api/v1/profiles/restore": {
      "post": {
        "tags": [
          "profiles"
        ],
        "summary": "Restores an export (SPEC-009 §2.1) into an empty installation. The same file again changes nothing; anything else stored is a 409.",
        "operationId": "restoreProfile",
        "requestBody": {
          "description": "An export file (SPEC-009 §2.1), at most 16777216 bytes.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileExport"
              }
            }
          },
          "required": true
        },
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestoredExport"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestoredExport"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "413": {
            "description": "Payload Too Large",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported Media Type",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKey": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ProfileExport": {
        "required": [
          "format",
          "formatVersion",
          "classification",
          "exportedAt",
          "entities"
        ],
        "type": "object",
        "properties": {
          "format": {
            "type": "string"
          },
          "formatVersion": {
            "type": "integer",
            "format": "int32"
          },
          "classification": {
            "type": "string"
          },
          "exportedAt": {
            "type": "string",
            "format": "date-time"
          },
          "entities": {
            "$ref": "#/components/schemas/ExportedEntities"
          }
        }
      },
      "ProfileInputDocument": {
        "required": [
          "taxYear",
          "region",
          "employment",
          "activity",
          "projection"
        ],
        "type": "object",
        "properties": {
          "taxYear": {
            "type": "integer",
            "format": "int32"
          },
          "region": {
            "type": "string"
          },
          "employment": {
            "$ref": "#/components/schemas/ProfileEmployment"
          },
          "activity": {
            "$ref": "#/components/schemas/ProfileRegistration"
          },
          "projection": {
            "$ref": "#/components/schemas/ProfileProjection"
          }
        },
        "additionalProperties": false
      },
...
      "ProfileView": {
        "required": [
          "id",
          "taxYear",
          "region",
          "employment",
          "activity",
          "projection"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "taxYear": {
            "type": "integer",
            "format": "int32"
          },
          "region": {
            "type": "string"
          },
          "employment": {
            "$ref": "#/components/schemas/ProfileEmployment"
          },
          "activity": {
            "$ref": "#/components/schemas/ProfileRegistration"
          },
          "projection": {
            "$ref": "#/components/schemas/ProfileProjection"
          }
        }
      },
...
    },
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "description": "The local installation's API key (SPEC-009 §3). The health checks do not need it.",
        "name": "X-Api-Key",
        "in": "header"
      }
    }
  },
...
}